src/requirements.txt
Django==1.8.6
django-markdown-deux==1.0.5
django-pagedown==0.1.3
django-taggit==0.22.1
markdown2==2.3.4
olefile==0.44
Pillow==4.1.1
pytz==2017.2
